Update Tags component
- Removed former Tags component
- Renamed TagList to Tags
- Rewrote Tags to accept tag limit, not character limit
- Rewrote stories and tests to match new Tags component
