+++ src/demo/sections/Tags.vue
@@ -1,15 +1,12 @@
 <template>
-<div>
-  <h3>Tags component with options</h3>
-  <div class="caption-small mb-2">(Renders optional <b>v-for</b> of <b>Tag</b>-s components, or returns "--" (2 hyphens) if a passed tags prop is an empty array. To style <b>Tags</b> pass same props (disabled, outlined, elevated etc.) as you would for <b>Tag</b>)</div>
-  <Tags :tags="this.tags" elevated disabled hovered/>
-  <br>
-  <Tags icon="pi-label" :tags="this.tags" flat color="primary"/>
-  <br>
-  <Tags icon="pi-fire-line" :tags="this.no_tags" elevated hovered/>
-  <br>
-  <Tags icon="pi-label" :tags="this.no_tags" flat/>
-</div>
+  <div>
+    <h3>Tags</h3>
+    <div class="caption-small mb-2">(Renders joint string of passed tags in flat style. If no tags passed, renders "--"(2 hyphens). If joint length of passed tags is over <b>tagLengthLimit</b> (can be passed as a prop, default 30) renders the size number of the array and <b>show</b> tooltip to see all the tags)</div>
+    <Tags icon="pi-prefect" :tags="[]" />
+    <Tags :tags="this.shortTags" />
+    <Tags icon="pi-earth-line" :tags="this.shortTags" :limit="2"/>
+    <Tags :tags="this.tags" />
+  </div>
 </template>
 
 <script lang="ts">
@@ -17,8 +14,8 @@ import { Vue, Options } from 'vue-class-component'
 
 @Options({})
 export default class Tags extends Vue {
-  tags = ["one", "two", "three"];
-  no_tags = []
+  tags = ["Apollo", "Orion", "Production", "Label", "Label2", "Label3"]
+  shortTags = ["Apollo", "Orion", "Production"]
 }
 </script>
